Abstract further; pull out client fetching into reusable hook.
